refinement

Struct Refinement

pub struct Refinement<T, P>(/* private fields */);

A Refinement type ensures all values of a particular type satisfy a Predicate.

Use as_inner/to_inner to access the underlying value or into_inner to unwrap the value.

Refinement also implements many common standard library traits if the underlying value also implements them.

Examples

use refinement::{Predicate, Refinement};

struct LessThanTen;

impl Predicate<i32> for LessThanTen {
    fn test(x: &i32) -> bool {
        *x < 10
    }
}

type LessThanTenInt = Refinement<i32, LessThanTen>;

let x = LessThanTenInt::new(5);
assert!(x.is_some());

let y = LessThanTenInt::new(11);
assert!(y.is_none());

Implementations

impl<T, P> Refinement<T, P>
where P: Predicate<T>,

pub fn new(x: T) -> Option<Self>

Create a refined value from the underlying type T.

Returns x under the refinement type if x satisfies P, otherwise returns None.

Examples

use refinement::{Predicate, Refinement};

struct NonEmpty;

impl Predicate<String> for NonEmpty {
    fn test(x: &String) -> bool {
       !x.is_empty()
    }
}

type NonEmptyString = Refinement<String, NonEmpty>;

let s1 = NonEmptyString::new(String::from("Hello"));
assert!(s1.is_some());

let s2 = NonEmptyString::new(String::from(""));
assert!(s2.is_none());

pub fn into_inner(self) -> T

Unwrap the underlying value, consuming self.

Examples

use refinement::{Predicate, Refinement};

struct ThreeDigit;

impl Predicate<String> for ThreeDigit {
    fn test(x: &String) -> bool {
       x.chars().count() == 3 && x.chars().filter(|c| c.is_ascii_digit()).count() == 3
    }
}

type ThreeDigitString = Refinement<String, ThreeDigit>;

let s = ThreeDigitString::new(String::from("123"));

assert_eq!(String::from("123"), s.unwrap().into_inner());

impl<T, P> Refinement<T, P>
where T: Clone, P: Predicate<T>,

pub fn to_inner(&self) -> T

Retrieve the underlying value without consuming self.

Examples

use refinement::{Predicate, Refinement};

struct ThreeDigit;

impl Predicate<String> for ThreeDigit {
    fn test(x: &String) -> bool {
       x.chars().count() == 3 && x.chars().filter(|c| c.is_ascii_digit()).count() == 3
    }
}

type ThreeDigitString = Refinement<String, ThreeDigit>;

let s = ThreeDigitString::new(String::from("123"));

assert_eq!(String::from("123"), s.unwrap().to_inner());

impl<T, P> Refinement<T, P>
where T: Copy, P: Predicate<T>,

pub fn as_inner(&self) -> T

Retrieve the underlying value for Copy types without consuming self.

Examples

use refinement::{Predicate, Refinement};

struct LessThanTen;

impl Predicate<i32> for LessThanTen {
    fn test(x: &i32) -> bool {
        *x < 10
    }
}

type LessThanTenInt = Refinement<i32, LessThanTen>;

let x = LessThanTenInt::new(5);
assert_eq!(5, x.unwrap().as_inner());